# perf: reuse result of BuildNewListFromBlock